requirements.txt
mysql-connector-python
fpdf
pandas
PyPDF2
streamlit_pdf_viewer
reportlab
